refactor(sensors): WIP refactor sensors/environmental_telemetry
Work on making both sensors and telemetry more board agnostic. Next:
make generic the I2C sensor device implementations of traits such that
this code can be used on both nrf and rp platforms. Probably will break
out into its own module/folder
